+++ src/lib/features/metrics/client-metrics/client-metrics-service.e2e.test.ts
@@ -1,15 +1,15 @@
-import ClientInstanceService from '../../../lib/services/client-metrics/instance-service';
-import { IClientApp } from '../../../lib/types/model';
+import ClientInstanceService from '../instance/instance-service';
+import { IClientApp } from '../../../types/model';
 import { secondsToMilliseconds } from 'date-fns';
-import { createTestConfig } from '../../config/test-config';
-import { IUnleashConfig, IUnleashStores } from '../../../lib/types';
-import { FakePrivateProjectChecker } from '../../../lib/features/private-project/fakePrivateProjectChecker';
-import { ITestDb } from '../helpers/database-init';
+import { createTestConfig } from '../../../../test/config/test-config';
+import { IUnleashConfig, IUnleashStores } from '../../../types';
+import { FakePrivateProjectChecker } from '../../private-project/fakePrivateProjectChecker';
+import { ITestDb } from '../../../../test/e2e/helpers/database-init';
 
 const faker = require('faker');
-const dbInit = require('../helpers/database-init');
-const getLogger = require('../../fixtures/no-logger');
-const { APPLICATION_CREATED } = require('../../../lib/types/events');
+const dbInit = require('../../../../test/e2e/helpers/database-init');
+const getLogger = require('../../../../test/fixtures/no-logger');
+const { APPLICATION_CREATED } = require('../../../types/events');
 
 let stores: IUnleashStores;
 let db: ITestDb;
