+++ src/lib/features/metrics/client-metrics/client-metrics-store-v2-type.ts
@@ -1,4 +1,4 @@
-import { Store } from './store';
+import { Store } from '../../../types/stores/store';
 
 export interface IClientMetricsEnvKey {
     featureName: string;
