+++ src/lib/features/metrics/client-metrics/client-metrics-store-v2.test.ts
@@ -1,6 +1,6 @@
-import dbInit, { ITestDb } from '../../test/e2e/helpers/database-init';
-import getLogger from '../../test/fixtures/no-logger';
-import { IClientMetricsStoreV2, IUnleashStores } from '../types';
+import dbInit, { ITestDb } from '../../../../test/e2e/helpers/database-init';
+import getLogger from '../../../../test/fixtures/no-logger';
+import { IClientMetricsStoreV2, IUnleashStores } from '../../../types';
 import { endOfDay, setHours, startOfHour, subDays } from 'date-fns';
 
 let stores: IUnleashStores;
