+++ src/lib/features/metrics/instance/register.test.ts
@@ -1,9 +1,9 @@
 import supertest from 'supertest';
-import { createTestConfig } from '../../../test/config/test-config';
-import createStores from '../../../test/fixtures/store';
-import getLogger from '../../../test/fixtures/no-logger';
-import getApp from '../../app';
-import { createServices } from '../../services';
+import { createTestConfig } from '../../../../test/config/test-config';
+import createStores from '../../../../test/fixtures/store';
+import getLogger from '../../../../test/fixtures/no-logger';
+import getApp from '../../../app';
+import { createServices } from '../../../services';
 
 async function getSetup() {
     const stores = createStores();
