+++ src/lib/features/metrics/last-seen/createLastSeenService.ts
@@ -1,5 +1,5 @@
-import FakeFeatureToggleStore from '../../../features/feature-toggle/fakes/fake-feature-toggle-store';
-import FeatureToggleStore from '../../../features/feature-toggle/feature-toggle-store';
+import FakeFeatureToggleStore from '../../feature-toggle/fakes/fake-feature-toggle-store';
+import FeatureToggleStore from '../../feature-toggle/feature-toggle-store';
 import { Db, IUnleashConfig } from '../../../server-impl';
 import { FakeLastSeenStore } from './fake-last-seen-store';
 import { LastSeenService } from './last-seen-service';
