+++ src/lib/features/metrics/last-seen/fake-last-seen-read-model.ts
