+++ src/lib/features/metrics/last-seen/last-seen-read-model.ts
