+++ src/lib/routes/client-api/index.ts
@@ -1,7 +1,7 @@
 import Controller from '../controller';
 import FeatureController from '../../features/client-feature-toggles/client-feature-toggle.controller';
-import MetricsController from './metrics';
-import RegisterController from './register';
+import MetricsController from '../../features/metrics/metrics';
+import RegisterController from '../../features/metrics/instance/register';
 import { IUnleashConfig, IUnleashServices } from '../../types';
 
 export default class ClientApi extends Controller {
