+++ src/lib/services/index.ts
@@ -5,8 +5,8 @@ import HealthService from './health-service';
 
 import ProjectService from './project-service';
 import StateService from './state-service';
-import ClientInstanceService from './client-metrics/instance-service';
-import ClientMetricsServiceV2 from './client-metrics/metrics-service-v2';
+import ClientInstanceService from '../features/metrics/instance/instance-service';
+import ClientMetricsServiceV2 from '../features/metrics/client-metrics/metrics-service-v2';
 import TagTypeService from '../features/tag-type/tag-type-service';
 import TagService from './tag-service';
 import StrategyService from './strategy-service';
@@ -35,7 +35,7 @@ import { ProxyService } from './proxy-service';
 import EdgeService from './edge-service';
 import PatService from './pat-service';
 import { PublicSignupTokenService } from './public-signup-token-service';
-import { LastSeenService } from './client-metrics/last-seen/last-seen-service';
+import { LastSeenService } from '../features/metrics/last-seen/last-seen-service';
 import { InstanceStatsService } from '../features/instance-stats/instance-stats-service';
 import { FavoritesService } from './favorites-service';
 import MaintenanceService from '../features/maintenance/maintenance-service';
@@ -85,7 +85,7 @@ import { FakeDependentFeaturesReadModel } from '../features/dependent-features/f
 import {
     createFakeLastSeenService,
     createLastSeenService,
-} from './client-metrics/last-seen/createLastSeenService';
+} from '../features/metrics/last-seen/createLastSeenService';
 import {
     createFakeGetProductionChanges,
     createGetProductionChanges,
