+++ src/lib/types/services.ts
@@ -5,7 +5,7 @@ import StateService from '../services/state-service';
 import StrategyService from '../services/strategy-service';
 import TagTypeService from '../features/tag-type/tag-type-service';
 import TagService from '../services/tag-service';
-import ClientInstanceService from '../services/client-metrics/instance-service';
+import ClientInstanceService from '../features/metrics/instance/instance-service';
 import ContextService from '../services/context-service';
 import VersionService from '../services/version-service';
 import { ApiTokenService } from '../services/api-token-service';
@@ -22,7 +22,7 @@ import FeatureToggleService from '../features/feature-toggle/feature-toggle-serv
 import EnvironmentService from '../features/project-environments/environment-service';
 import FeatureTagService from '../services/feature-tag-service';
 import ProjectHealthService from '../services/project-health-service';
-import ClientMetricsServiceV2 from '../services/client-metrics/metrics-service-v2';
+import ClientMetricsServiceV2 from '../features/metrics/client-metrics/metrics-service-v2';
 import UserSplashService from '../services/user-splash-service';
 import { OpenApiService } from '../services/openapi-service';
 import { ClientSpecService } from '../services/client-spec-service';
@@ -32,7 +32,7 @@ import { ProxyService } from '../services/proxy-service';
 import EdgeService from '../services/edge-service';
 import PatService from '../services/pat-service';
 import { PublicSignupTokenService } from '../services/public-signup-token-service';
-import { LastSeenService } from '../services/client-metrics/last-seen/last-seen-service';
+import { LastSeenService } from '../features/metrics/last-seen/last-seen-service';
 import { InstanceStatsService } from '../features/instance-stats/instance-stats-service';
 import { FavoritesService } from '../services/favorites-service';
 import MaintenanceService from '../features/maintenance/maintenance-service';
